Refactor request store and improve dashboard metrics
Replaced RequestStore with a new MemoryRequestStore implementing IRequestStore for async log handling. Updated dependency injection and middleware to use the new store. Refactored NucleusRequestLogService and NucleusRequestFlushService for improved batching and async streaming. Enhanced dashboard UI with new chart styles and real-time metrics, and migrated TESTaPP to MiniOps.Nucleus.TestApi with updated configuration files and project references.

+++ MiniOps.Nucleus.Core/Config/NucleusOptions.cs
@@ -101,7 +101,7 @@ public sealed class NucleusOptions
     /// </summary>
     /// <param name="interval">Flush interval in seconds.</param>
     /// <returns>The current <see cref="NucleusOptions"/> instance for fluent configuration.</returns>
-    public NucleusOptions SetBatchFlushInterval(int interval)
+    public NucleusOptions SetBatchFlushInterval(float interval)
     {
         BatchFlushIntervalSeconds = interval;
         return this;
